Refactored resources and tests directory
