+++ guard/resources/test-command/data-dir/s3_bucket_logging_enabled_tests.yaml
