+++ guard/resources/test-command/data-dir/s3_bucket_server_side_encryption_enabled.yaml
