+++ guard/resources/validate/data-dir/s3-public-read-prohibited-template-compliant.yaml
