+++ guard/resources/validate/data-dir/s3-public-read-prohibited-template-non-compliant.yaml
