+++ guard/resources/validate/data-dir/s3-server-side-encryption-template-non-compliant.yaml
