+++ guard/resources/validate/db_resource.yaml
