+++ guard/resources/validate/input-parameters-dir/db_params.yaml
