+++ guard/resources/validate/input-parameters-dir/dummy.txt
