+++ guard/resources/validate/malformed-rule.guard
