+++ guard/resources/validate/malformed-template.yaml
