+++ guard/resources/validate/output-dir/test_single_data_file_single_rules_file_verbose.out
