+++ guard/resources/validate/rules-dir/dummy.txt
