+++ guard/resources/validate/rules-dir/s3_bucket_logging_enabled.guard
