+++ guard/resources/validate/rules-dir/s3_bucket_public_read_prohibited.guard
