+++ guard/resources/validate/rules-dir/s3_bucket_server_side_encryption_enabled.guard
