+++ guard/resources/validate/s3-server-side-encryption-template-non-compliant-2.yaml
