+++ guard/resources/validate/s3_bucket_server_side_encryption_enabled_2.guard
