+++ guard/tests/utils.rs
@@ -9,6 +9,9 @@ use std::path::PathBuf;
 use cfn_guard::command::Command;
 use cfn_guard::utils::writer::{WriteBuffer, Writer};
 use clap::App;
+use cfn_guard::commands::{migrate::Migrate,test::Test,validate::Validate,
+                          rulegen::Rulegen, parse_tree::ParseTree};
+
 
 pub fn read_from_resource_file(path: &str) -> String {
     let mut resource = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
@@ -44,51 +47,7 @@ pub(crate) fn compare_write_buffer_with_string(
     assert_eq!(expected_output, actual_output)
 }
 
-pub fn cfn_guard_test_command<T: Command>(command: T, args: Vec<&str>) -> i32 {
-    let test_app_name = "cfn-guard-test";
-    let mut app = App::new(test_app_name);
-    let mut command_options = Vec::new();
-    command_options.push(test_app_name);
-    command_options.append(args.clone().as_mut());
-
-    let mut commands: Vec<Box<dyn Command>> = Vec::with_capacity(2);
-    commands.push(Box::new(command));
-
-    let mappings = commands.iter().map(|s| (s.name(), s)).fold(
-        HashMap::with_capacity(commands.len()),
-        |mut map, entry| {
-            map.insert(entry.0, entry.1.as_ref());
-            map
-        },
-    );
-
-    for each in &commands {
-        app = app.subcommand(each.command());
-    }
-
-    let app = app.get_matches_from(command_options);
-
-    match app.subcommand() {
-        (name, Some(value)) => {
-            if let Some(command) = mappings.get(name) {
-                match (*command).execute(value, &mut Writer::new(WriteBuffer::Stdout(stdout()))) {
-                    Err(e) => {
-                        println!("Error occurred {}", e);
-                        -1
-                    }
-                    Ok(code) => code,
-                }
-            } else {
-                -2
-            }
-        }
-
-        (_, None) => -3,
-    }
-}
-
-pub fn cfn_guard_test_command_verbose<T: Command>(
-    command: T,
+pub fn cfn_guard_test_command(
     args: Vec<&str>,
     mut writer: &mut Writer,
 ) -> i32 {
@@ -99,7 +58,11 @@ pub fn cfn_guard_test_command_verbose<T: Command>(
     command_options.append(args.clone().as_mut());
 
     let mut commands: Vec<Box<dyn Command>> = Vec::with_capacity(2);
-    commands.push(Box::new(command));
+    commands.push(Box::new(Validate::new()));
+    commands.push(Box::new(Test::new()));
+    commands.push(Box::new(ParseTree::new()));
+    commands.push(Box::new(Migrate::new()));
+    commands.push(Box::new(Rulegen::new()));
 
     let mappings = commands.iter().map(|s| (s.name(), s)).fold(
         HashMap::with_capacity(commands.len()),
